+++ Products/CMFPlone/tests/testCatalogTool.py
@@ -9,6 +9,7 @@ from plone.indexer.wrapper import IndexableObjectWrapper
 from plone.uuid.interfaces import IAttributeUUID
 from plone.uuid.interfaces import IUUID
 from Products.CMFCore.permissions import AccessInactivePortalContent
+from Products.CMFCore.indexing import processQueue
 from Products.CMFPlone.CatalogTool import CatalogTool
 from Products.CMFPlone.CatalogTool import is_folderish
 from Products.CMFPlone.tests import dummy
@@ -154,6 +155,7 @@ class TestCatalogIndexing(PloneTestCase):
         self.folder.invokeFactory('Document', id='doc',
                                   title='Foo', description='Bar')
         self.catalog.unindexObject(self.folder.doc)
+        processQueue()
 
     def assertResults(self, result, expect):
         # Verifies ids of catalog results against expected ids
@@ -214,6 +216,7 @@ class TestCatalogIndexing(PloneTestCase):
     def testReindexObjectSkipsMetadata(self):
         # Reindexing should not update metadata when update_metadata=0
         self.catalog.indexObject(self.folder.doc)
+        processQueue()
         self.folder.doc.setTitle('Fred')
         self.folder.doc.setDescription('BamBam')
         self.catalog.reindexObject(self.folder.doc, update_metadata=0)
@@ -226,6 +229,7 @@ class TestCatalogIndexing(PloneTestCase):
     def testReindexTitleOnly(self):
         # Reindexing should only index the Title
         self.catalog.indexObject(self.folder.doc)
+        processQueue()
         self.folder.doc.setTitle('Fred')
         self.folder.doc.setDescription('BamBam')
         self.catalog.reindexObject(self.folder.doc, idxs=['Title'])
@@ -249,6 +253,7 @@ class TestCatalogIndexing(PloneTestCase):
     def testReindexTitleOnlySkipsMetadata(self):
         # Reindexing Title should not update metadata when update_metadata=0
         self.catalog.indexObject(self.folder.doc)
+        processQueue()
         self.folder.doc.setTitle('Fred')
         self.folder.doc.setDescription('BamBam')
         self.catalog.reindexObject(self.folder.doc, idxs=['Title'],
@@ -267,6 +272,7 @@ class TestCatalogIndexing(PloneTestCase):
         # by searchResults()!?
         #
         self.catalog.indexObject(self.folder.doc, idxs=['Title'])
+        processQueue()
         # The document is cataloged
         path = self.catalog._CatalogTool__url(self.folder.doc)
         self.assertTrue(path in self.catalog._catalog.paths.values())
@@ -282,6 +288,7 @@ class TestCatalogIndexing(PloneTestCase):
         # of index type.
         #
         self.catalog.indexObject(self.folder.doc, idxs=['getId'])
+        processQueue()
         # The document is cataloged
         path = self.catalog._CatalogTool__url(self.folder.doc)
         self.assertTrue(path in self.catalog._catalog.paths.values())