+++ CHANGES.rst
@@ -14,7 +14,9 @@ Breaking changes:
 
 New features:
 
-- *add item here*
+- Adapt code and tests to the new indexing operations queueing.
+  Part of PLIP 1343: https://github.com/plone/Products.CMFPlone/issues/1343
+  [gforcada]
 
 Bug fixes: